Fix bug in barrier implementation
When copying the implementation of `ParallelContext::thread_barrier`, I
included the definition of `myCycle` as `static thread_local` and its
initialization to 0.
The corresponding `_barrier_proceed` was made a class member variable,
and is therefore initialized to 0 on every `PartitionModelEvaluation`.
When an uneven number of calls to barrier is made, `myCycle` might still be
set to 1 when the evaluation is finished, causing a mismatch when
switching to a new `PartitionModelEvaluation` where `_barrier_proceed` is
initialized as 0.

The fix is to reset `myCycle` whenever a thread joins a new
`PartitionModelEvaluation`.